requirements.txt
# Core dependencies
psycopg2-binary>=2.9.3
openai>=1.1.0
apify-client>=1.0.0
python-dotenv>=1.0.0
schedule>=1.2.0

# RAG dependencies
torch>=2.0.0
transformers>=4.30.0
langchain>=0.0.267
langchain-openai>=0.0.1
langchain-community>=0.0.1
pymilvus>=2.2.16
pillow>=10.0.0

# Email dependencies
Jinja2>=3.1.2

# Utility libraries
requests>=2.31.0
tqdm>=4.65.0

# Optional: Development Tools
black>=23.3.0  # Code formatting
pylint>=2.17.0  # Code linting
numpy==1.26.2
